Final documentation and readme updates
Fixed mistake in dockerfile where port 8000 was being exposed instead of 5000

Complete overhaul to documentation to ensure that the project is clearly outlined in a literal roadmap.

Added a flowchart which visualises what the project has achieved, and how it operates, Update to README.md to annotate this to provide further understanding to what's been completed, why and how.